+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 
 ### Added
 - `sites list` now has an optional `--cached` flag which makes the command return the cached sites data rather than retrieving it anew. (#637)
+- `sites mass-update` now can be filtered by `--tag=<tag>`. (Note: `--org=<name|id>` is necessary to use the filter.) (#640)
+- `sites mass-update` now has an optional `--cached` tag to optionally prevent retrieving a new sites cache. (#640)
 - Environment variables `TERMINUS_PORT` and `TERMINUS_PROTOCOL` now enabled. (#643)
 
 ##[0.9.1] - 2015-10-27